+++ src/lib.rs
@@ -478,6 +478,7 @@ mod header;
 pub mod svq1_blocktree;
 pub mod svq1_codebook;
 pub mod svq1_helper_luts;
+pub mod svq1_mc;
 pub mod svq1_mean;
 pub mod svq1_motion_predictor;
 pub mod svq1_mv_cache;